Problem: sequence flows are unconditional
This makes it impossible to build any logic in BPMN. However, BPMN does
have conditional sequence flows, it's just this implementation of BPXE
doesn't have them at this early stage.

Solution: make sequence flows conditional

Firstly, I had to amend the schema generator to account for the type
substitution trick used in BPMN to supply formal expressions in place of
any expression (via `xsi:type` attribute) by introducing a wrapper type
(`AnExpression`) that can be either an `Expression` or a
`FormalExpression`.

Then, since expressions are text nodes, I had to amend all structures
generated by the schema generator to contain this text (accessible
through `Element`'s `TextPayload()`).

Then, I added a mapping from expression language URIs to their
respective engines (`expression.GetEngine`).

Now, all that had to be done was to add expression evaluation to
`flow.Flow` handling of outgoing flows and filter out sequence flows
that don't flow. I had to deal with the case when neither of them flows,
and added a special trace called `FlowTerminationTrace` (though it could
have been `FlowTrace` with `SequenceFlows` set to an empty slice -- but
at the time I decided it's better to emphasize the fact that flow has
been terminated to avoid any misinterpretation; we'll see how well will
this age).

In the course of doing the above work, I realized `Tracer` had a number
of bugs with regards to subscription and unsubscription that led to
deadlocks. That has been fixed. I really should add some `Tracer`
specific tests.

Keep in mind that the usefulness of conditional sequence flows is still
very limited. Or, more precisely, they are virtually useless as they are
completely isolated from any data sources at this point. This is to be
addressed as one of the next steps.