Problem: inability to execute any logic
This (again) severly limits the ability of BPXE to do anything useful.

Solution: implement Exclusive Gateway

This allows workflows to make logical decisions, effectively enabling
rudimentary flow control.

There's a bit to unpack in this commit.

The interface for flow_node.Outgoing has changed a bit since we need to
be able to have "conversations" with flows between multiple calls. For
that, flows have gotten unique identifiers (through introduction of
`bpxe.org/pkg/id` package).

There was also a bug in `bpmn.AnExpression` unmarshaller. It was
expecting "tFormalExpression" as a value, but that would only be a case
if bpmn namespace was the default namespace. Now we check if
"tFormalExpression" is a suffix in a form of ":tFormalExpression" or if
it is "tFormalExpression" exactly. It is not perfect as we're not using
actual bpmn namespace name, but I've found no way to do this with Go's
`encoding/xml` interface.

This also implements sequence flow probing in flows so it can be used in
other gateways for their needs.