Problem: permissive licensing
Building and maintaining the project comes with a significant cost.

Solution: re-license under the terms of BSL 1.1

In order to fund its development, BPXE is hereby switching to Business
Source License. This allows for keeping source code available for the
entire product, permits non-commercial uses (through Additional Use
Grant) and releases it under the terms of Apache 2.0 on a scheduled basis
(4 years past the release), preserving it for future, more permissive
use.