Problem: no way to react to events
This severely limits what BPXE is able to accomplish.

Solution: implement intermediateCatchEvent and eventBasedGateway

Intermediate Catch Events are now able to continue the flow when defined
events occur. It is important to note that this change DOES NOT implement
functionality for actual observation of events (such as timer or conditional events)
but rather exposes a generic mechanism for observing these. To this end, `event.Instance`
and `event.InstanceBuilder` were added to facilitate model/process/process instance-specified
creation of such event instances. (I don't particular like calling these things
Event Instances, but I couldn't come up with a better name just yet)

This change does not implement eventDefinitionRefs as per specification yet mostly
due to the fact that BPMN schema was not yet generated 100% properly to support abstract elements
resolving to other abstract elements (event definition in root elements). Once this is improved,
Intermediate Catch Events should be able to handle this rather easily.

Please note that `pkg/events` was renamed to `pkg/event`.

Event-Based Gateways commanded an addition of two features to flow actions (specifically, to
FlowAction): ActionTransformer and Terminate. The former allows to capture the next action and
transform it, the latter allows to terminate (cancel) flows that are no longer of interest.
I don't particularly *love* this current design, but it's something I can iterate from.

This change does not implement Receive Task support for Event-Based Gateways at this moment.
This is something to address separately to limit the scope of the change.

This change amends schema generator to support returning a slice of an interface type for
abstract elements. The schema generator also improves reliability of generated FindBy implementation
to ensure it checks if the receiver is in fact non-nil.