Problem: explaining open source nature of BPXE
README says BPXE is not Open Source software (according to OSI, at the
very least). However, the goal is to make it so, but there's a cost to
making software, so BPXE takes a gradual approach to achieving its full
open source status, while retaining the important benefits of source
availability, hackability and personal + non-commercial use grant.

Solution: update wording in README to reflect this