Problem: parallel gateway join & fork
Parallel gateway works fine when it's either joining or forking,
but not both.

Solution: map sequence flows to pending flow actions properly

This also fixes a bug in EndEvent when its runner was exiting once it
was sending `NoAction` or `CompleteAction` back to flow. This caused subsequent
flows to lock up as there was nobody to handle these requests.