Problem: intermediate_catch_event_test still occasionally times out
Solution: fix a race condition in IntermediateCatchEvent

There was a problem in how IntermediateCatchEvent was handling Incoming &
NextAction. It was activated upon Incoming but only added pending Action
channels upon getting NextAction. If a matching event is received between
these two steps, they will be successfully processed but there will be no
Action channels to report to. Effectively, these matched events would have
no observable effects on the flow.

Similar change is made to EventBasedGateway. EventBasedGateway can be used
to instantiate a process (not supported by BPXE just yet) and therefore it
won't have an Incoming in this case.