Problem: event/catch.New using `intermediateCatchEvent`
This parameter name is confusing. Are we limited to
intermediate catch events here?

Solution: rename it to `catchEvent`

This naming is a relic of the past when this code was
for intermediate catch events only. As you can see in the type of
the parameter, all `CatchEvent`s are supported.