Problem: intermittent test timeouts
This primarily happens in boundary_test.go

Solution: ensure observation of listening trace

The boundary test assumes that a boundary event is already
listening post-instantiation of the process. However,
this is not necessarily true, because this is timing-dependant
as all the flow nodes are independent goroutines.

So, we need to wait until a trace indicating boundary event is
listening has been received. This check used to be in this test
but was erroneously removed when it didn't work as expected.

The problem was that the test subscribed to the tracer too late --
at a point when listening trace might have been sent out as process
instantiation provisions activity harness which starts boundary flows
and respective traces are emitted.

This hinges on the way tracers are assumed to be created. They are created
upon process instantiation. There's no way to create a tracer before
the process is instantiated. This change fixes this oversight. It adds
`process.InstanceOption` type that allows to modify instance right after
it is constructed by `process.NewInstance`. First such option is `WithTracer`
that'll set a custom tracer in the instance.

With this change, the tests can now create their own tracer ahead of time
and pass it to instances.

To celebrate tests no longer timing out (fingers crossed!) I'm upping test count
for test runs to 100 to have better assurances every time we run the tests.
This required me to decrease timeouts in some of the tests so that they finish
sooner. I hope they can be improved upon later to ensure `time.After` doesn't
need to be used for testing at all. I've also ensured license check test
runs only once because it doesn't need to be ran more than once and it's pretty
slow.