Problem: test runs are a bit slow
Now that we run 100 of each tests in each invocation of
`make test`, use of timeouts in tests has become a bit of a burden
as the testing iterations became longer.

Solution: make waiting for timeouts unnecessary

Two new traces were introduced:

`event/catch.EventObservedTrace` to signal that `event/catch.Node`
has observed a particular event. This allowed us to make sure that
the events have or have not reached the event catcher

`gateway/parallel.IncomingFlowProcessedTrace` to signal that an incoming
flow has been processed (so, any outgoing flow already happened)

These traces allowed me to rewrite tests in such a way that the expectation
of timeouts is no longer necessary.

I've also broke out license testing from `count=100` run so that it can be
fully cached between invocations of `make test`. I've not done this for CI
as there's still a check in that test to ensure it runs only once.