Problem: inclusive gateway tests hanging
These timeouts are somewhat unpredictable, but they
happen fairly regularly.

Solution: re-engineer parts of inclusive gateway

Firstly, some assurances were made in the flow tracker. Particularly,
it will now attempt to drain the traces in full before allowing
`activeFlowsInCohort` to return a list of active flows in the cohort (via
the use of the lock); nor will it attempt sending an activity notification
until it's drained the traces. There was one particular situation where
the flow tracker was locking up: it was attempting to send an activity notification
while holding the lock. This meant that if the inclusive gateway node was
calling `activeFlowsInCohort` and waiting for the lock there, the flow tracker
itself would block on attempting to send an activity notification. Increasing
its buffer helps to a degree, but its hard to predict the size of the queue
necessary. Either way, sending multiple notifications for a single drain doesn't
seem to make a lot of sense.

While working on that, I also wanted to make sure that the tracker is indeed
always "up to speed": it will now lock up at the construction time and will
not unlock until it has observed the flow that comes into the node.

Secondly, inclusive gateway was redesigned to make tracking of flows that it is
waiting on a bit easier to reason about. Initially, it was deleting elements from
`awaiting` list when they came in, and this meant that if that list was to change
due to updates from the tracker, there could be an inconsistent view of what's
happening. Now it's tracking arrived flows separately and compares them with the
list of what it is awaiting for.

Thirdly, the test was updated to use its own tracker to avoid missing out on
anything.

However, I was still experiencing intermittent timeouts in tests. So I dug deeper
and I discovered that occasionally traces come in a wrong order (and they were
definitely not supposed to do that!). The culprit was in the `flow` package:
`handleAdditionalSequenceFlow` eagerly started new flows before the current
flow had a chance to log the flow, resulting in a wrong order. This issue
has been addressed.

Note: I've observed a slowdown of inclusive gateway after this update. I think
it's about 2.5 times slower now. Not great, but perhaps it is something that can
be addressed down the road and optimized.