Problem: Incoming interface in flow_node
While it's being used, there doesn't seem to be any
good reason why it should exist in addition to NextAction.
It was carried over as a design from the previous iteration of
BPXE implementation, but not much thought has been given to it.
I feel like it unnecessarily complicates the code.

An index of NextAction can be figured by comparing `flow.SequenceFlow()`
of the flow to those of incoming flows of the node.

Solution: remove it and simplify code

If there will ever be a reason for something like this, we can
cross this bridge when we get there.