Problem: absence of data flow
The usability of BPXE is still pretty limited as it can't
store or retrieve any data during process workflow.

Solution: sketch initial support for data objects

A newly introduced `data` package is responsible for working
with data. Container type is responsible for holding and providing
access to instances of ItemAware elements.

Speaking of which, the schema generator was extended to introduce ItemAware
and ItemAwareElement abstract types to match BPMN specification. For this to
happen, a modified XSchema specification was used to allow multiple type
extensions. Since this is not done for the purposes of schema validation,
but rather to enable code generation, I think this is acceptable.

This change also fixes an issue in the schema generator that
added `TextPayload` fields to structs of abstract types. This was
tripping up `xml.Decoder` as it was seeing a conflict of the payload
between the abstract type and the specific one.

Sequence flows now support accessing data objects through `getDataObject`
function with following caveats:

* Only a single-argument version of `getDataObject` is currently supported
  (two-argument version allows to refer to a parent process; since we don't
   support this yet, this has not been implemented)
* XPath engine does not yet support `getDataObject`. Some work has been done
  in this direction, but the test for this is currently being skipped as it
  doesn't work. The problem is (seemingly) in not being able to access the
  NodeSet returned by a function using the following path. I am trying to figure
  out if this has something to do with how xsel work and whether it can be
  made to work in this case.
* `getDataObject` does not currently distinguish between data objects, data
  object references and properties. This should be addressed to make it more
  precise.

The event subsystem is precipitated with the ability for certain messages to
carry data -- their structures have been augmented to support that. However,
this change does not implement this just yet. It is just a rough sketch to
work on in smaller, incremental fashion.