Problem: flow nodes require too many parameters
<flow_node>.New() methods are taking a huge number of
parameters and this hurts readability.

Solution: pass all this information in one struct

Most of these parameters are universally shared and were passed
through flow node implementations to `flow_node.New`.

Now, that structure is constructed prior to creation of a particular
flow node and a reference to it is passed to the flow node,
cutting down on a number of arguments required.

To celebrate this improvement, `flow_node.T` was renamed to a much
more fitting `flow_node.Wiring` (at least it's more informative than `T`)