Problem: explicit process instantiation only
Currently, a process can be only instantiated explicitly, but
this is not very useful for situations when processes need to be
triggered.

Solution: start implementing further instantiation semantics
from BPMN specification.

This is a fairly big change. Firstly, events have been extracted
from process scope into a model scope (where they should have belonged
in the first place, in fact) and renamed from `ProcessEvent` to `Event`.

Secondly, a running model (one started by `model/Model.Run`) will
now register event consumers that will look for events defined in
flow nodes that are eligible for process instantiation (see
`bpmn/Process.InstantiatingFlowNodes`) and will instantiate a process
if events do match. It's currently only implemented for Start Events
and needs to be extended to Event-Based Gateways as well as Receive Tasks
(which are not yet implemented).

`process/Instance` have been extracted to `process/instance/Instance` mostly
to separate processes and instances logically, but also to allow the same
names for `WithXXX` constructor options to allow passage of contexts and
other information.

It is important to note here that the way event egress and ingress is done
for flow nodes has changed a bit: instance itself is now supposed to
consume events from the model and forward them to flow nodes; flow nodes, however,
send out their events to the model. This allows to specify different entry
points for events. So, publishing into an instance will only send an event
into its subscribers; publishing into a model will send it to all subscribers.