Problem: usage of event.Instance
Firstly, it's a confusing term because there's a process
instance.

Secondly, events seem to type assert their way to `definitionDistance`,
which defies the purpose of the interface

Solution: rename event.Instance to event.DefinitionInstance
and add EventDefinition() to the interface that the code can use.

It's still using the word "instance" (can't think of anything better)
but at least it's using it as `DefinitionInstance` which would hopefully
separate it from `process/instance/Instance`