Problem: timer events are sent manually
For the time being, all timer events are being sent
manually in tests. This is not enough for a functioning
BPMN engine.

Solution: introduce `timer` package

This is something to be used by the engine initiate timer
events. It's not yet integrated into the engine.

Speaking of the mock time implementation, and picking
https://github.com/benbjohnson/clock.

It is known to have issues and is not actively maintained,
here's a short overview:

https://github.com/gofrs/help-requests/issues/35

There are other alternatives, and I tried a few of them,
but for some reason they don't seem to work for the tests I wrote:

https://github.com/derision-test/glock
https://github.com/microsoft/clock
https://github.com/tilinna/clock

I haven't tried the approach laid out in https://github.com/cabify/timex
yet.

My current hypothesis is that it doesn't really matter which one
will be picked as long as it facilitates proper testing, as they
all delegate to `time` package for non-mocking use cases. That being said,
the tests of `timer` package are a bit on the slow side, and that's a nuisance.