Problem: current mock clock testing is slow
I am not 100% sure why it is slow, but it is slow,
taking 2.5 seconds on Linux and about 30 seconds on Windows.

Solution: implement BPXE's own clock abstraction
to exercise full control over it.

While at it, I modelled an important addition to its interface -
`Until` function that returns a channel that it will send a message
to once a certain time is reached, unlike `After` which forces
timer to calculate the duration even if it has a set time to wait for.
This has simplified the design of timer considerably.

It also implements a `Changes()` channel that will notify the timer
every time there's a change or a drift. I think it's useful functionality,
even though I ended up ripping the usage of it out of timers since they
now use `Until()` to achieve what they need to achieve. Yet, I decided to
keep it for the future needs. The implementation of it for the real clock
was extracted from `timer.DriftingClock` I did before.

The tests seem to run approximately 100 times faster on Linux now. Hopefully,
they have a similarly improved performance profile on Windows.