Problem: host clock drift checking
The algorithm used is a bit naive and purely observational,
doesn't bring a lot of trust.

Solution: introduce Linux-specific mechanism

The idea is based on using a Linux-specific timerfd API and
create a timer with an expiration at the very end of absolute
time and TFD_TIMER_CANCEL_ON_SET flag set, which means the
timer will be cancel if time changes:

"If this flag is specified along with TFD_TIMER_ABSTIME and
the clock for this timer is CLOCK_REALTIME or
CLOCK_REALTIME_ALARM, then mark this timer as cancelable
if the real-time clock undergoes a discontinuous change
(settimeofday(2), clock_settime(2), or similar).  When
such changes occur, a current or future read(2) from the
file descriptor will fail with the error ECANCELED."

This, combined with non-blocking reader from this timer
file descriptor once a second is the basis of this Linux-specific
mechanism.

If there's a good way to do this for other platforms, this
kind of platform-specific code should be introduced as well.