Problem: timer is not being used by the engine
This renders its usefulness at zero, as timer catch events
still don't function.

Solution: integrate timer into the event system

This changed the interface for event definition instances
a bit (mostly to be able to return errors).

Clock got an ability to be retrieved from context (though I am not sure
I like this very much as this is not very explicit -- perhaps
WithClock would be better -- as a usability improvement).