Problem: BPXE is developed by a single person
It means that my own biases, myopia and experience stand
in the way of producing the best quality software reasonably
possible. Having a bigger team (more eyes) working on this
would have been helpful, but we're not there yet.

Solution: extend an open invitation for code reviews

In order to facilitate such reviews, provide incentives
in a form of monetary compensation and other perks, if
possible.