Problem: impossible clock logic in host's change monitor
There's a test that tests whether a certain time is before
itself. That's clearly impossible.

Solution: check whether it's before previously known time

While we're at it, update current time on every iteration of the loop.

The reason for this mishap is that this piece of code is not
well tested yet (since it can be only tested manually)

Addresses #64