Problem: tracking mutex unlocking patterns
Some of the code has branching with its own unlocking
of mutexes. This code is error-prone as it's easy to
forget to unlock.

Solution: use deferred unlocks where feasible

Previous implementation was based on a now-outdated knowledge
that deferrals are slower. As it was pointed out in #68, this is
no longer true as of Go 1.14. My own measurements seem to confirm
this as well.

So, where it was reasonable, I've rewritten unlocking to happen in
deferred statements.