Problem: timer tests fail sporadically
The error seems to indicate that something
is being received from the channel when it's
not expected to.

Solution: handle channel closure

Since timers close channels when it is safe to do so,
sometimes tests receive another (zero) message that
indicates channel closure.

This change handles this behavior by observing optional
flag that indicates whether the channel is closed.