Problem: process instance observing traces from outside
Currently, there's no reason for process instance components to observe
traces from outside of its own scope; in fact, it'll likely mistreat
traces from other instances or other processes as its own.

Solution: compartmentalize tracers

This way, instance gets its own tracer and its tracer are relayed to
process's tracer and so on. This removes transforming tracer as it is
no longer necessary. Instead, tracer relay does its job.

I am still not 100% confident this is the right approach but it is better
than leaving things as they are with scopes bleeding into other
scopes. I am trying to think if there's any reason for observing traces
outside of tracing compartment. If there will be such a need, we'll
need to wire it through accordingly.